+++ rector.php
@@ -14,8 +14,8 @@ return static function (RectorConfig $rectorConfig): void {
 
     // define sets of rules
     $rectorConfig->sets([
-        LevelSetList::UP_TO_PHP_71,
-        SymfonySetList::SYMFONY_34,
+        LevelSetList::UP_TO_PHP_74,
+        SymfonySetList::SYMFONY_44,
         SymfonySetList::SYMFONY_CODE_QUALITY,
         SymfonySetList::SYMFONY_CONSTRUCTOR_INJECTION,
     ]);