+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+                                                        //
+## [Unreleased]
+
+### Changed
+
+- The minimum PHP version is now 8.1.  Applied Rector-suggested changes.
+- The minimum Symfony version is now 4.4.
 
 ## [1.0.0] - 2023-03-31
 ### Changed